requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
cryptography==41.0.3
dotenv==0.0.5
firebase-admin==6.2.0
Flask==2.3.2
Flask-Cors==4.0.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.5
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
plotly==5.17.0
pocketsphinx==5.0.2
proto-plus==1.22.3
protobuf==4.24.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.1.1
pypiwin32==223
pypng==0.20220715.0
python-dotenv==1.0.0
pyttsx3==2.90
pywin32==306
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.0
tenacity==8.2.3
tweepy==4.14.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.7
